+++ Projects/AccountSDKIOSWeb/generate_mocks.sh
@@ -1,7 +1,6 @@
 ./cuckoo/run generate --testable AccountSDKIOSWeb \
   --output Tests/AccountSDKIOSWebTests/GeneratedMocks.swift \
     ../../Sources/AccountSDKIOSWeb/Lib/HTTP/HTTPClient.swift \
-    ../../Sources/AccountSDKIOSWeb/Lib/HTTP/URLSessionProtocol.swift \
     ../../Sources/AccountSDKIOSWeb/Lib/Storage/SessionStorage.swift \
     ../../Sources/AccountSDKIOSWeb/Lib/Storage/Storage.swift \
     ../../Sources/AccountSDKIOSWeb/Lib/Storage/Keychain/Compat/LegacyKeychainSessionStorage.swift \
